+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/trie/pathbased/common/provider/PathBasedWorldStateProvider.java
@@ -132,12 +132,6 @@ public abstract class PathBasedWorldStateProvider implements WorldStateArchive {
             Bytes32.wrap(rootHash.getBytes()), blockHash);
   }
 
-  @Override
-  public boolean isWorldStateImmediatelyCached(final Hash blockHash) {
-    return worldStateCacheManager.contains(blockHash)
-        || headWorldState.blockHash().equals(blockHash);
-  }
-
   /**
    * Gets a mutable world state based on the provided query parameters.
    *